Move SumTest files. Fix bug in LargeDenseRowBlock reset function
