+++ src/test/scripts/functions/aggregate/TensorSum.dml
